chore(docs): Optimized array indexing in the log monitor widget's rendering logic
# Project State

## Current Focus
Optimized array indexing in the log monitor widget's rendering logic

## Context
The change improves performance by removing unnecessary type casting during array indexing in the log monitor widget's rendering logic.

## Completed
- [x] Removed redundant `as usize` cast in array index calculation

## In Progress
- [x] Performance optimization for log monitor widget rendering

## Blockers
- None identified

## Next Steps
1. Verify performance impact with larger log datasets
2. Review other potential indexing optimizations in related widgets
---
category: chore
scope: docs
---

+++ examples/_cookbook/log_monitor.rs
@@ -224,7 +224,7 @@ impl Widget for LogMonitor {
         let content_bottom = h - 2;
         for y in content_top..content_bottom {
             for x in 1..w - 1 {
-                let idx = (y * w + x) as usize;
+                let idx = y * w + x;
                 if idx < p.cells.len() {
                     p.cells[idx].bg = t.surface;
                 }
